ARM: Tegra: modify caching technique in gpu-edp
Change the data structure used to cache EDP-constrained frequencies in
gpu-edp.c. The old data structure was an array of frequencies indexed
by temperature and having a fixed EDP budget (in mA). The new data
structure is a hash table which caches fmax for multiple EDP budgets
and temperatures. The purpose of the change is to support both fixed
EDP limits (a la regulator EDP management) and variable EDP limits (a
la sysedp) with the same code.

Change-Id: I8c93065d0842c133e6cd98aa2edfb8d3dd3c4e7a
Signed-off-by: Matt Longnecker <mlongnecker@nvidia.com>
Reviewed-on: http://git-master/r/496384
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Timo Alho <talho@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>