lowmemorykiller: don't consider CMA memory as free
As CMA memory can only be used for Movable allocations,
Android lowmemkiller shouldn't consider this as free memory
for lowmem kill decisions.  If it considered as free memory,
It can delay the killing of apps and lead to slowness of
system as memory allocations start waiting for kswap
daemon to free up the memory.

Bug 1550455

Change-Id: Ica8da48ef48912d89cde526f0ba3676df6ab10f1
Signed-off-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-on: http://git-master/r/552250