video: tegra: nvmap: create per-pid debugfs memory data
/sys/kernel/debug/nvmap/handles_by_pid/<pid> will list the nvmap handles
owned by that pid in a versioned binary format.  This is intended for
automated collection of memory usage by a specific process.  For this
usecase, human-readability and per-heap breakdowns are unimportant, and
the reduced allocation size can help reduce memory fragmentation.

For the curious, /sys/kernel/debug/nvmap/<heap>/maps will still contain
per-heap mapping data in a human-readable format.

Bug: 17689644

Change-Id: I22544a449c000e4bfec549cfadbb165703b64bd3
Signed-off-by: Greg Hackmann <ghackmann@google.com>