WAR: flounder: remove 204Mhz emc dvfs entry
This is a workaround for audio glitch and freeze issue with some streaming video apps.
We found that after device suspend/resume cycle, the problem may occur. The DMA operation may get stuck so that alsa ring buffer has no space for app to write data. We also found that bumping up emc frequency will eliminate this problem. This WAR is to step over certain emc frequency and bump it up to a higer level.

Note that this is a WAR instead of a fix. It may impact power consumption. We don't know why it works yet and we are keeping investigating.

Bug: 18158174
Change-Id: I6780d95a6dc7e080c9352a406f2ed5eb42d7c519