mm: vmscan: fix wakeup_kswapd's per zone disable logic
While wakeup_kswapd's first parameter represents the current zone
for which allocation logic is trying to wake kswapd for, the
per zone wake up disable logic was using just the preferred zone's
idx to prevent kswapd from wake up.
This is incorrect. Consider a case where preferred zone is ZONE_HIGHMEM
but current zone is ZONE_NORMAL. Ideally this logic should be bypassed
as request is for ZONE_NORMAL but it is not.

So, even though there is a need for kswapd to reclaim memory for an
allocation, it does not do so. This has serious side effects and in some
cases hang as well.

zonelist describes the allowed zones for the current allocation. Allow
scan on only the lowest zone in the zonelist to fix logic and avoid such
incorrect behaviour.

Bug 200130931
Bug 200105279

Change-Id: Iaca11ccf0404dd16ebb9266c64b40293b99eb96e
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/839001
Reviewed-on: http://git-master/r/936109
(cherry picked from commit 084b2ddd8c0a107fdb9618853b742d7607d69efb)
Reviewed-on: http://git-master/r/1000491
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>