Lower threshold for embedded memory controller clock transitions
For very low-load animations, the EMC clock would be reduced to
the point that the render threads consumed 100% of the CPU for
more than 20ms, triggering a CPU clock transition to the
"hispeed_freq" of 1.5ghz. The CPU clocks would eventually settle
back down to the minimum and it would happen again.

Reduce the EMC threshold to eliminate this clock thrashing.

BUG: 27880019
Change-Id: Iac55a2ec1732958a45d1209a58538a0bca0745b3