Revert "ANDROID: flounder: don't steal pages from CMA"
This is the one that caused the full system freeze
that happened after some uptime.

But what I found interesting is that the stock kernel does not seem to
have the problem (Same with ElementalX kernel). And to add to my confusion
just building stock kernel will have the freeze GNU/GOOGLE gcc.

This reverts commit c644d303929af77a7e3281a21a4ff26cbe88179a.

Signed-off-by: USBhost <josiahspore@gmail.com>