Makefile: correct cache values; add all graphite flags.
It turns out I have been compiling with the wrong l1 cache values for a long time now. It always seemed to bug me so after I had enough. I did some research to find out I was using the values of the two cores when I was only supposed to be using one core's instruction cache (l1i). So its 64 where it should have been to begin with not 196 (l1i+l1d*2). About the l2 cache values IDK how it got to be 2097 so its back to 2048 where it belongs.

Signed-off-by: USBhost <josiahspore@gmail.com>